Added a command line player static class
Basically this will be the command line tester Ryan wanted for the
cooker. See the bottom of program.cs to see how to run use it in the
tester. You will need to make a function in server game that calls the
CommandLinePlayer function with the gameObjects as a list. The
serverGame has it set to private.